Refactor: extract chunk test setup to new package chunk/testutils
